+++ Mailer/extensions/mailer-module/src/Models/Generators/ArticleUrlParserGenerator.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace Remp\MailerModule\Models\Generators;
 
 use Nette\Application\UI\Form;
+use Nette\Http\Url;
 use Nette\Utils\ArrayHash;
 use Remp\Mailer\Components\GeneratorWidgets\Widgets\ArticleUrlParserWidget\ArticleUrlParserWidget;
 use Remp\MailerModule\Models\ContentGenerator\Engine\EngineFactory;
@@ -16,11 +17,18 @@ class ArticleUrlParserGenerator implements IGenerator
 {
     public $onSubmit;
 
+    private array $contentProcessors = [];
+
     public function __construct(
         protected SourceTemplatesRepository $sourceTemplatesRepository,
         protected ContentInterface $content,
-        private EngineFactory $engineFactory
+        protected readonly EngineFactory $engineFactory,
     ) {
+    }
+
+    public function setContentProcessor(string $domain, $contentProcessor)
+    {
+        $this->contentProcessors[$domain] = $contentProcessor;
     }
 
     public function generateForm(Form $form): void
@@ -30,11 +38,22 @@ class ArticleUrlParserGenerator implements IGenerator
             ->setHtmlAttribute('class', 'form-control wysiwyg-editor')
             ->getControlPrototype();
 
-        $form->addTextArea('articles', 'Article')
+        $form->addTextArea('articles', 'Articles')
             ->setHtmlAttribute('rows', 7)
             ->setHtmlAttribute('class', 'form-control')
             ->setOption('description', 'Paste article URLs. Each on separate line.')
             ->setRequired(true)
+            ->getControlPrototype();
+
+        $form->addTextArea('external_intro', 'External intro text')
+            ->setHtmlAttribute('rows', 4)
+            ->setHtmlAttribute('class', 'form-control wysiwyg-editor')
+            ->getControlPrototype();
+
+        $form->addTextArea('external_articles', 'External articles')
+            ->setHtmlAttribute('rows', 7)
+            ->setHtmlAttribute('class', 'form-control')
+            ->setOption('description', 'Paste article URLs from external website supported by the generator. Each on separate line.')
             ->getControlPrototype();
 
         $form->addTextArea('footer', 'Footer text')
@@ -73,6 +92,8 @@ class ArticleUrlParserGenerator implements IGenerator
             (new PostInputParam('footer'))->setRequired(),
             (new PostInputParam('utm_campaign'))->setRequired(),
             (new PostInputParam('intro'))->setRequired(),
+            (new PostInputParam('external_intro')),
+            (new PostInputParam('external_articles')),
         ];
     }
 
@@ -84,32 +105,19 @@ class ArticleUrlParserGenerator implements IGenerator
             throw new \RuntimeException("Unable to find source template with ID [{$values['source_template_id']}]");
         }
 
-        $items = [];
         $errors = [];
-
         $urls = explode("\n", trim($values['articles']));
-        foreach ($urls as $url) {
-            $url = trim($url);
-            if (empty($url)) {
-                // people sometimes enter blank lines
-                continue;
-            }
-            try {
-                $meta = $this->content->fetchUrlMeta($url);
-                if ($meta) {
-                    $items[$url] = $meta;
-                } else {
-                    $errors[] = $url;
-                }
-            } catch (InvalidUrlException $e) {
-                $errors[] = $url;
-            }
-        }
+        $items = $this->readArticlesMetadata($urls, $errors);
+
+        $externalUrls = explode("\n", trim($values['external_articles']));
+        $externalItems = $this->readArticlesMetadata($externalUrls, $errors);
 
         $params = [
             'intro' => $values['intro'],
+            'external_intro' => $values['external_intro'],
             'footer' => $values['footer'],
             'items' => $items,
+            'external_items' => $externalItems,
         ];
 
         $engine = $this->engineFactory->engine();
@@ -128,5 +136,37 @@ class ArticleUrlParserGenerator implements IGenerator
     public function preprocessParameters($data): ?ArrayHash
     {
         return null;
+    }
+
+    private function readArticlesMetadata(array $urls, &$errors): array
+    {
+        $items = [];
+
+        foreach ($urls as $url) {
+            $url = trim($url);
+            if (empty($url)) {
+                // people sometimes enter blank lines
+                continue;
+            }
+            try {
+                $parsedUrl = new Url($url);
+                if (isset($this->contentProcessors[$parsedUrl->getDomain()])) {
+                    $processor = $this->contentProcessors[$parsedUrl->getDomain()];
+                    $meta = $processor->fetchUrlMeta($url);
+                } else {
+                    $meta = $this->content->fetchUrlMeta($url);
+                }
+
+                if ($meta) {
+                    $items[$url] = $meta;
+                } else {
+                    $errors[] = $url;
+                }
+            } catch (InvalidUrlException $e) {
+                $errors[] = $url;
+            }
+        }
+
+        return $items;
     }
 }